runpod-ipadapter/requirements.txt
runpod==1.3.3
diffusers==0.25.0
transformers==4.36.0
accelerate==0.25.0
safetensors==0.4.1
torch==2.1.1
torchvision==0.16.1
Pillow==10.1.0

numpy==1.26.4
huggingface-hub==0.19.4
einops==0.6.1
xformers==0.0.23

# FaceID dependencies
insightface==0.7.3
onnxruntime-gpu>=1.16.0

# YOLOv8 + SAM for advanced segmentation
ultralytics>=8.0.0
segment-anything>=1.0
opencv-python-headless>=4.8.0

# ControlNet dependencies
controlnet-aux>=0.0.7
